pyproject.toml
[project]
name = "code-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.4.0",
    "openai>=1.59.6",
    "requests>=2.32.3",
]
